# Remove an allocation in pipeline processing