@mediocregopher 2021-01-23
Well damn, and here I had thought I was being efficient by _not_ allocating that error each loop. Thanks!